% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ERA_NWP.R
\name{get_ERA_NWP}
\alias{get_ERA_NWP}
\title{get_ERA_NWP}
\usage{
get_ERA_NWP(
  ERA_path = "~/Desktop/MasterNR/PC_ERA_79_92.Rda",
  NWP_path = "~/Desktop/MasterNR/NWP_quant_1993_2022.Rda",
  NWP_quant = "NWP_q_90_PC1",
  PC = "PC1",
  NWP_preds = 1
)
}
\arguments{
\item{ERA_path}{Path to ERA file}

\item{NWP_path}{Path to NWP file}

\item{NWP_quant}{NWP quantile of interest}

\item{PC}{Principle component}

\item{NWP_preds}{Number of NWP predictors}
}
\value{
Data.table with data ready for input in rolling_mod function
}
\description{
get_ERA_NWP
}
\examples{
ERA_NWP = get_ERA_NWP(NWP_preds = 2)


}
