% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forc_data_re.R
\name{get_forc_data_re}
\alias{get_forc_data_re}
\title{Get Reweighted Forecast Data
This is a cover function where one inputs the range of forecasts months.
The function then iterates over the specified range calling get_quantiles_reweighted,
and stores the results of these iterated outputs in a data.table.
Output can then be merged with ERA-data using a get_ERA_NWP function to be used in prediction tasks.}
\usage{
get_forc_data_re(start_month, start_year, forc_months)
}
\arguments{
\item{forc_months}{}
}
\value{
Returns reweighted quantiles of interest
}
\description{
Get Reweighted Forecast Data
This is a cover function where one inputs the range of forecasts months.
The function then iterates over the specified range calling get_quantiles_reweighted,
and stores the results of these iterated outputs in a data.table.
Output can then be merged with ERA-data using a get_ERA_NWP function to be used in prediction tasks.
}
