% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinball_loss.R
\name{pinball_loss}
\alias{pinball_loss}
\title{pinball loss}
\usage{
pinball_loss(quantile, pred, obs)
}
\arguments{
\item{quantile}{Quantile of interest}

\item{pred}{Predicted values}

\item{obs}{Observed values}
}
\value{
Loss
}
\description{
pinball loss
}
\examples{
pinball_loss(0.9, 2, 2)

}
