% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_climat_rolling.R
\name{rolling_mod_climatology}
\alias{rolling_mod_climatology}
\title{rolling_mod_climatology}
\usage{
rolling_mod_climatology(q, ERA_NWP, window = 60)
}
\arguments{
\item{q}{Quantile of interest.}

\item{ERA_NWP}{Data of ERA observations and quantiles of NWP forecasts.}

\item{window}{Max number of forecast days from initialization.}
}
\value{
Model output with predictions, loss and model coefficients.
}
\description{
rolling_mod_climatology
}
