% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_mod_re.R
\name{rolling_mod_re}
\alias{rolling_mod_re}
\title{Rolling model reweighting}
\usage{
rolling_mod_re(
  q,
  ERA_NWP_re = load("input_data/ERA_NWP.Rda"),
  predictors,
  model = "reg",
  window = 45,
  hour_v = FALSE,
  week_v = FALSE,
  month_v = FALSE,
  year_v = FALSE
)
}
\arguments{
\item{q}{Quantile of interest.}

\item{predictors}{Number of NWP predictors.}

\item{model}{Specify type of model, e.g. qreg, qgam, etc.}

\item{window}{Max number of forecast days from initialization.}

\item{hour_v}{Include hour covariate.}

\item{week_v}{Include week covariate.}

\item{month_v}{Include month covariate.}

\item{year_v}{Include year covariate.}

\item{ERA_NWP}{Data of ERA observations and quantiles of NWP forecasts.}
}
\value{
Model output with predictions, loss and model coefficients.
}
\description{
Rolling model reweighting
}
